auth_project/requirements.txt
amqp==2.6.0
asgiref==3.2.10
billiard==3.6.3.0
celery==4.4.6
Django==3.0.8
django-ranged-response==0.2.0
django-redis==4.12.1
django-simple-captcha==0.5.12
dnspython==2.0.0
eventlet==0.25.2
future==0.18.2
greenlet==0.4.16
importlib-metadata==1.7.0
itsdangerous==1.1.0
kombu==4.6.11
monotonic==1.5
msgpack==1.0.0
msgpack-python==0.5.6
Pillow==7.2.0
pytz==2020.1
redis==3.5.3
six==1.15.0
sqlparse==0.3.1
vine==1.3.0
zipp==3.1.0
